# docs/api_swagger.yaml
basePath: /api
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  schema.Company:
    properties:
      address:
        type: string
      code:
        type: string
      createdAt:
        type: string
      delete:
        type: boolean
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      domain:
        type: string
      fax:
        type: string
      id:
        type: string
      name:
        type: string
      phone_number:
        type: string
      sector:
        type: string
      status:
        type: boolean
      updatedAt:
        type: string
    type: object
...
  schema.Division:
    properties:
      code:
        type: string
      createdAt:
        type: string
      delete:
        type: boolean
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      estate:
        $ref: '#/definitions/schema.Estate'
      estate_code:
        type: string
      gangs:
        items:
          $ref: '#/definitions/schema.Gang'
        type: array
      id:
        type: string
      name:
        type: string
      status:
        type: boolean
      updatedAt:
        type: string
    type: object
...
  schema.Estate:
    properties:
      code:
        type: string
      company:
        $ref: '#/definitions/schema.Company'
      company_code:
        type: string
      createdAt:
        type: string
      delete:
        type: boolean
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      divisions:
        items:
          $ref: '#/definitions/schema.Division'
        type: array
      id:
        type: string
      name:
        type: string
      status:
        type: boolean
      updatedAt:
        type: string
    type: object
  schema.EstateResponse:
    properties:
      Division:
        items:
          $ref: '#/definitions/schema.Division'
        type: array
      code:
        type: string
      company:
        $ref: '#/definitions/schema.Company'
      delete:
        type: boolean
      description:
        type: string
      id:
        type: string
      name:
        type: string
      status:
        type: boolean
    type: object
  schema.Gang:
    properties:
      code:
        type: string
      delete:
        type: boolean
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      division:
        $ref: '#/definitions/schema.Division'
      division_code:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        type: boolean
    type: object
...
  schema.JsonEmployeeCreate:
    properties:
      address:
        type: string
      application:
        items:
          type: string
        type: array
      code:
        type: string
      company_id:
        type: string
      department:
        type: string
      division:
        items:
          type: string
        type: array
      email:
        type: string
      fullname:
        type: string
      nickname:
        type: string
      nik:
        type: string
      office_number:
        type: string
      password:
        type: string
      phone_number:
        type: string
      picture:
        type: string
      role:
        items:
          type: string
        type: array
      role_application:
        items:
          type: string
        type: array
      username:
        type: string
    required:
    - code
    - email
    - password
    - username
    type: object
  schema.JsonEmployeeUpdate:
    properties:
      address:
        type: string
      application:
        items:
          type: string
        type: array
      code:
        type: string
      company_id:
        type: string
      department:
        type: string
      division:
        items:
          type: string
        type: array
      email:
        type: string
      fullname:
        type: string
      nickname:
        type: string
      nik:
        type: string
      office_number:
        type: string
      password:
        type: string
      phone_number:
        type: string
      picture:
        type: string
      role:
        items:
          type: string
        type: array
      role_application:
        items:
          type: string
        type: array
      user_id:
        type: string
      username:
        type: string
    required:
    - code
    - email
    - username
    type: object
...
  schema.MetaCompany:
    properties:
      address:
        type: string
      code:
        type: string
      description:
        type: string
      domain:
        type: string
      fax:
        type: string
      id:
        type: string
      name:
        type: string
      phoneNumber:
        type: string
      sector:
        type: string
    type: object
...
  schema.MetaEmployee:
    properties:
      address:
        type: string
      code:
        type: string
      company:
        $ref: '#/definitions/schema.MetaCompany'
      companyCode:
        type: string
      costCenter:
        type: string
      department:
        type: string
      email:
        type: string
      employeeType:
        type: string
      expired:
        type: integer
      fiscalPeriod:
        type: string
      fiscalYear:
        type: string
      fullName:
        type: string
      gradeRate:
        type: string
      harvesterCode:
        type: string
      id:
        type: string
      jobPos:
        type: string
      mandt:
        type: string
      nickName:
        type: string
      nik:
        type: string
      officeNumber:
        type: string
      phoneNumber:
        type: string
      picture:
        type: string
      userCode:
        type: string
      username:
        type: string
    type: object
...
  schema.MsgResponse:
    properties:
      msg:
        type: string
    type: object
...
  schema.RequestCompanyUpdate:
    properties:
      address:
        type: string
      code:
        type: string
      delete:
        type: boolean
      description:
        type: string
      domain:
        type: string
      fax:
        type: string
      name:
        type: string
      phone_number:
        type: string
      sector:
        type: string
      status:
        type: boolean
    required:
    - address
    - code
    - delete
    - description
    - domain
    - fax
    - name
    - phone_number
    - sector
    - status
    type: object
...
  schema.RequestEstate:
    properties:
      code:
        type: string
      company_code:
        type: string
      description:
        type: string
      name:
        type: string
    required:
    - code
    - company_code
    - description
    - name
    type: object
  schema.RequestEstateUpdate:
    properties:
      code:
        type: string
      company_code:
        type: string
      delete:
        type: boolean
      description:
        type: string
      name:
        type: string
      status:
        type: boolean
    required:
    - code
    - company_code
    - delete
    - description
    - name
    - status
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: Fajar soegi
    url: http://www.swagger.io/support
  description: This is User Service Test.
  termsOfService: http://swagger.io/terms/
  title: Swagger User Service
  version: "1.0"
paths:
...
  /employee:
    post:
      consumes:
      - application/json
      description: Employee Create New
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 'Employee Add New Schema '
        in: body
        name: Register
        required: true
        schema:
          $ref: '#/definitions/schema.JsonEmployeeCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.MsgResponse'
      summary: Employee Add New
      tags:
      - Employee
  /employee/{id}:
    get:
      consumes:
      - application/json
      description: Get Meta User By ID or Code
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: EMPLOYEE ID OR EMPLOYEE Code
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.MetaEmployee'
      summary: Get Meta User By ID or Code
      tags:
      - Employee
  /employee/{id}/edit:
    patch:
      consumes:
      - application/json
      description: Employee Update Field
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: EMPLOYEE ID OR EMPLOYEE Code
        in: path
        name: id
        required: true
        type: string
      - description: 'Employee Update '
        in: body
        name: Register
        required: true
        schema:
          $ref: '#/definitions/schema.JsonEmployeeUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.MsgResponse'
      summary: Employee Update
      tags:
      - Employee
  /estate:
    get:
      consumes:
      - application/json
      description: Get Estate
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/schema.EstateResponse'
            type: array
      summary: Get Estate
      tags:
      - (B) Estate
    patch:
      consumes:
      - application/json
      description: Update Company
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 'Company Schema '
        in: body
        name: Estate
        required: true
        schema:
          $ref: '#/definitions/schema.RequestCompanyUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.MsgResponse'
      summary: Update Company
      tags:
      - (A) Company
    post:
      consumes:
      - application/json
      description: Create Estate
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 'Estate Schema '
        in: body
        name: Estate
        required: true
        schema:
          $ref: '#/definitions/schema.RequestEstate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.MsgResponse'
      summary: Create Estate
      tags:
      - (B) Estate
  /estate/{id}:
    get:
      consumes:
      - application/json
      description: Get Estate ID
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Estate ID OR Estate Code
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.EstateResponse'
      summary: Get Estate By ID
      tags:
      - (B) Estate
  /estate/{id}/edit:
    patch:
      consumes:
      - application/json
      description: Update Estate
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Estate ID OR Estate Code
        in: path
        name: id
        required: true
        type: string
      - description: 'Estate Schema '
        in: body
        name: Estate
        required: true
        schema:
          $ref: '#/definitions/schema.RequestEstateUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.MsgResponse'
      summary: Update Field Estate
      tags:
      - (B) Estate
...
swagger: "2.0"